Add electrum-ltc support

+++ pkgs/electrum-like.nix
@@ -12,13 +12,16 @@ let
       --add-flags "-D" \
       --add-flags '"''${XDG_DATA_HOME:-$HOME/.local/share}/${bin}"' \
 
-    rm $out/share/applications/${bin}.desktop
-    sed -E 's#${pkg}#'"$out"'#' \
-      ${pkg}/share/applications/${bin}.desktop \
-      > $out/share/applications/${bin}.desktop \
+    if [[ -e $out/share/applications/${bin}.desktop ]]; then
+      rm $out/share/applications/${bin}.desktop
+      sed -E 's#${pkg}#'"$out"'#' \
+        ${pkg}/share/applications/${bin}.desktop \
+        > $out/share/applications/${bin}.desktop
+    fi
   '';
 in
 {
   electrum = wrap prev.electrum "electrum";
+  electrum-ltc = wrap prev.electrum-ltc "electrum-ltc";
   electron-cash = wrap prev.electron-cash "electron-cash";
 }
